refs #23 starte merging the "insert" and "attach" systems into a more cohesive coupling/organization
