+++ src/anime/mod.rs
@@ -1,6 +1,4 @@
 pub mod anime_component;
 pub mod anime_res;
-pub mod attach_animations;
-pub mod insert_animations;
-pub mod overlay_animations;
-pub mod tile_animations;
+pub mod character_anime_sys;
+pub mod map_anime_sys;
