+++ src/main.rs
@@ -16,26 +16,20 @@ use yakuzaishi::{NINTENDO_DS_SCREEN_HEIGHT, NINTENDO_DS_SCREEN_WIDTH};
 use yakuzaishi::anime::anime_res::{
     AnimationAssets, EnvironmentEntityAnimationAssets, PlayerEntityAnimationAssets,
 };
-use yakuzaishi::anime::attach_animations::{
-    attach_animations_to_environment_entities, attach_animations_to_individual_tile_entities,
-    attach_animations_to_player_entities, attach_base_textures_to_player_entities,
+use yakuzaishi::anime::character_anime_sys::{
+    animate_env_entity_animations, attach_animations_to_environment_entities,
+    attach_animations_to_player_entities, insert_overlay_animation_resources_into_world,
 };
-use yakuzaishi::anime::insert_animations::{
-    insert_overlay_animation_resources_into_world, insert_tile_animation_resources_into_world,
-};
-use yakuzaishi::anime::overlay_animations::animate_env_entity_animations;
-use yakuzaishi::anime::tile_animations::{
+use yakuzaishi::anime::map_anime_sys::{
     animate_overlapped_tiles_event_based, handle_overlap_event, TileAnimationEvent,
 };
 use yakuzaishi::audio::audio_res::AudioAssets;
 use yakuzaishi::audio::audio_sys::start_background_audio;
-use yakuzaishi::camera::camera::track_camera;
-use yakuzaishi::camera::camera_load::init_camera;
-use yakuzaishi::environment::environment_load::spawn_environment_entity;
-use yakuzaishi::map::tiled::{TiledLoader, TiledMap, TiledMapAssets};
-use yakuzaishi::map::tiled_map::{process_tiled_maps, spawn_tiled_map_entity};
-use yakuzaishi::player::player_load::spawn_player_entity;
-use yakuzaishi::player::player_run::control_player_entity;
+use yakuzaishi::camera::camera_sys::{init_camera, track_camera};
+use yakuzaishi::environment::environment_sys::spawn_environment_entity;
+use yakuzaishi::map::tiled_res::{TiledLoader, TiledMap, TiledMapAssets};
+use yakuzaishi::map::tiled_sys::{process_tiled_maps, spawn_tiled_map_entity};
+use yakuzaishi::player::player_sys::{control_player_entity, spawn_player_entity};
 
 fn main() {
     let _tracy_client = Client::start();
@@ -75,7 +69,6 @@ fn main() {
             (
                 start_background_audio,
                 spawn_tiled_map_entity, // TODO: I don't like the name of this because its spawning an asset dependant thing which i feel like should be called Load
-                insert_tile_animation_resources_into_world,
                 insert_overlay_animation_resources_into_world,
                 spawn_player_entity,
                 spawn_environment_entity,
@@ -87,10 +80,8 @@ fn main() {
             OnExit(GameState::AssetProcessing),
             (
                 process_tiled_maps,
-                attach_animations_to_individual_tile_entities,
-                attach_base_textures_to_player_entities,
-                attach_animations_to_player_entities, // TODO: REQUIRED OnExit
-                attach_animations_to_environment_entities, // TODO: REQUIRED OnExit
+                attach_animations_to_player_entities,
+                attach_animations_to_environment_entities,
             ),
         )
         .add_event::<TileAnimationEvent>()
