+++ src/player/mod.rs
@@ -1,2 +1,1 @@
-pub mod player_load;
-pub mod player_run;
+pub mod player_sys;
